Move directive tests to another directory
